Data Dig Project
This project purpose was to extract necessary data one would conceivably need from a table such as this. This project forced me to ask questions that one would only ask in a position that required this information. This project, as well as learning SQL has given me an opportunity to not only learn this new language, but to ask necessary questions about how to sort and organize large databases.